# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Minecraft Log Observatory API",
    "version": "0.1.0",
    "description": "Local API for generated Minecraft log reports, split report store metadata, refresh jobs, and rule-editor support."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8787"
    }
  ],
  "paths": {
...
    "/api/app/status": {
      "get": {
        "summary": "First-run, setup state with nextActions, recovery actions, launcher contract, project, report/store readiness including missing store files, refresh, and refresh-needed status",
        "responses": {
          "200": {
            "$ref": "#/components/responses/JsonResponse"
          },
          "400": {
            "$ref": "#/components/responses/JsonResponse"
          },
          "500": {
            "$ref": "#/components/responses/JsonResponse"
          }
        }
      }
    },
...
    "/api/rules/test": {
      "post": {
        "summary": "Classify one raw chat message using current rules",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RuleTestRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/JsonResponse"
          },
          "400": {
            "$ref": "#/components/responses/JsonResponse"
          }
        }
      }
    },
...
  },
  "components": {
    "responses": {
      "JsonResponse": {
        "description": "JSON response",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiEnvelope"
            }
          }
        }
      }
    },
    "schemas": {
      "ApiEnvelope": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean"
          },
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "additionalProperties": true
      },
...
      "RuleTestRequest": {
        "type": "object",
        "required": [
          "message"
        ],
        "properties": {
          "message": {
            "type": "string"
          },
          "ruleSets": {
            "type": "array",
            "description": "Optional known rule set ids to limit matching for this test request.",
            "items": {
              "type": "string"
            }
          },
          "customRulePaths": {
            "type": "array",
            "description": "Optional safe project-relative rule-pack directories or .json files used only for this test request.",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
    }
  }
}